Backend refactor (#13)
* Backend refactor

This original goal of this commit was to return the full state of the cashout
and bids for a given game when a player cashes out. This was unfortunately
non-trivial given the current organization of the backend. I've refactored
things a bit to better suit this task, and hopefully other ones later on.

The rough idea of the refactor is to have a single dict in the game class
to handle the player tracking. The mapping is now WebSocket -> PlayingPlayer.
The PlayingPlayer instances track name, bids, cashout mult and cashout gain.

* Fix small bug in cashout logic