Asyncio + small db changes (#18)
## Asyncio
Replace threaded logic with asyncio.
Everything now runs on the same thread in a concurrent but not parallel fashion.

## DB
Load all data from db and instantiate a PlayingPlayer with it. Use a single instance of player in all places; get rid of other types representing player. The historical data (bids, mults, cashouts) are all read and re-written at each transaction.